+++ src/zepben/evolve/streaming/get/network_consumer.py
@@ -529,6 +529,10 @@ class NetworkConsumerClient(CimConsumerClient[NetworkService]):
             to_resolve = set()
             for obj in res.objects:
                 for ref in self.service.get_unresolved_references_from(obj):
+                    # Skip any reference trying to resolve from an EquipmentContainer - e.g a PowerTransformer trying to pull in its LvFeeder.
+                    # EquipmentContainers should be retrieved explicitly or via a hierarchy call.
+                    if EquipmentContainer in ref.resolver.from_class.__bases__:
+                        continue
                     to_resolve.add(ref.to_mrid)
 
             response = await self._get_identified_objects(to_resolve)